refactor: deck 패키지 분리
