+++ src/test/java/blackjack/model/PlayingDeckTest.java
@@ -4,6 +4,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import blackjack.model.card.Card;
+import blackjack.model.deck.DeckGenerator;
+import blackjack.model.deck.PlayingDeck;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
